requirements/base.txt
# Data Handling
pandas==1.5.3
numpy==1.23.5
scipy==1.10.1

# Machine Learning
scikit-learn==1.0.2
xgboost==1.6.2
imbalanced-learn==0.10.1
joblib==1.2.0

# Feature Engineering
feature-engine==1.6.1
category-encoders==2.6.1

# Model Interpretability
shap==0.41.0


python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors